requirements.txt
os
crewai
crewai-tools
langchain-openai
python-docx
python-dotenv